ref: Add trimming processor
